Moved requests package to outer file and directory changed
