+++ tests/app_aws/requirements.txt
@@ -1,2 +1,1 @@
 boto3[crt]
-requests
