--- tests/web_service/requirement.txt
@@ -1,1 +1,0 @@
-requests==2.26.0
